+++ qiskit_class_converter/services/converter_service.py
@@ -21,16 +21,17 @@ converter service
 import typing
 from enum import Enum
 
+import numpy as np
 from qiskit import QuantumCircuit
 
-from qiskit_class_converter.converters.braket_notation_to_matrix \
-    import BraketNotationToMatrixConverter
 from qiskit_class_converter.converters.matrix_to_quantum_circuit \
     import MatrixToQuantumCircuitConverter
 from qiskit_class_converter.converters.quantum_circuit_to_braket_notation \
     import QuantumCircuitToBraketNotationConverter
 from qiskit_class_converter.converters.quantum_circuit_to_matrix \
     import QuantumCircuitToMatrixConverter
+from qiskit_class_converter.converters.string_to_braket_notation \
+    import StringToBraketNotationConverter
 
 
 class ConversionType(Enum):
@@ -76,14 +77,17 @@ result = sample_converter.convert(input_value=input_value)
 quantum_circuit = QuantumCircuit(2, 2)
 quantum_circuit.append(result, [0, 1])
 ```"""
-    BRA_KET_TO_MATRIX = BraketNotationToMatrixConverter
+    STR_TO_BRA_KET = StringToBraketNotationConverter
     """```python
 from qiskit import QuantumCircuit
 from qiskit_class_converter import ConversionService
 
-sample_converter = ConversionService(conversion_type="BRA_KET_TO_MATRIX")
+sample_converter = ConversionService(conversion_type="STR_TO_BRA_KET")
 sample_converter.convert(input_value="|01>")
 ```"""
+    BRA_KET_TO_MATRIX = StringToBraketNotationConverter
+    """.. warning:: BRA_KET_TO_MATRIX was changed to STR_TO_BRA_KET and deprecated. \
+    This naming will be removed in 0.2 and later."""
 
 
 class ConversionService:  # pylint: disable=too-few-public-methods
@@ -100,7 +104,7 @@ ConversionService(conversion_type="QC_TO_BRA_KET", option={"expression": "simpli
     def __init__(self, conversion_type: typing.Union[str, ConversionType], option=None):
         """
         init function
-        :param conversion_type:  QC_TO_BRA_KET, QC_TO_MATRIX, MATRIX_TO_QC, BRA_KET_TO_MATRIX
+        :param conversion_type:  QC_TO_BRA_KET, QC_TO_MATRIX, MATRIX_TO_QC, STR_TO_BRA_KET
         :param option: See the Options table in this article.
         """
         if option is None:
@@ -111,7 +115,7 @@ ConversionService(conversion_type="QC_TO_BRA_KET", option={"expression": "simpli
         elif isinstance(conversion_type, ConversionType):
             self.__conversion_object = conversion_type.value
 
-    def convert(self, input_value: typing.Union[list, QuantumCircuit, str]):
+    def convert(self, input_value: typing.Union[list, np.ndarray, QuantumCircuit, str]):
         """
         convert functions